Added filing validators class, rearranged code
