+++ src/sbl_filing_api/config.py
@@ -93,6 +93,8 @@ class RequestActionValidations(BaseSettings):
         "check_contact_info",
     }
 
+    filing_create: Set[str] = {"check_period_exists", "check_period_filing_exists"}
+
     model_config = SettingsConfigDict(env_prefix="request_validators__", env_file=env_files_to_load, extra="allow")
